@@ -1,52 +1,52 @@
-require File.join(File.dirname(__FILE__), '..', 'spec_helper.rb')
+require 'spec_helper.rb'
 
-describe DummyGammaController do  
+RSpec.describe DummyGammaController, type: :controller do
   describe "for handling requests for HTML" do
     it "should use controller defined methods to handle login required redirection when available" do
       get :index
-      response.status.should == 200
+      expect(response).to be_success
     end
-  
+
     it "should use controller defined methods to handle permission denied redirection when available" do
       get :index
-      response.status.should == 200
+      expect(response).to be_success
     end
   end
 
   describe "for handling requests for XML" do
     it "should use controller defined methods to handle login required redirection when available" do
       get :index, :format => "xml"
-      response.status.should == 200
+      expect(response).to be_success
     end
-  
+
     it "should use controller defined methods to handle permission denied redirection when available" do
       get :index, :format => "xml"
-      response.status.should == 200
+      expect(response).to be_success
     end
   end
 
   describe "for handling requests for JSON" do
     it "should use controller defined methods to handle login required redirection when available" do
       get :index, :format => "json"
-      response.status.should == 200
+      expect(response).to be_success
     end
-  
+
     it "should use controller defined methods to handle permission denied redirection when available" do
       get :index, :format => "json"
-      response.status.should == 200
+      expect(response).to be_success
     end
   end
-  
+
   describe "for handling requests for JS" do
     it "should use controller defined methods to handle login required redirection when available" do
       get :index, :format => "js"
-      response.status.should == 200
+      expect(response).to be_success
     end
-  
+
     it "should use controller defined methods to handle permission denied redirection when available" do
       get :index, :format => "js"
-      response.status.should == 200
+      expect(response).to be_success
     end
   end
-  
+
 end
