@@ -54,6 +54,6 @@ Dummy::Application.routes.draw do
 
   # This is a legacy wild controller route that's not recommended for RESTful applications.
   # Note: This route will make all actions in every controller accessible via GET requests.
-  match ':controller(/:action(.:format))'
-  match ':controller(/:action(/:id(.:format)))'
+  get ':controller(/:action(.:format))'
+  get ':controller(/:action(/:id(.:format)))'
 end
